requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
aplus==0.11.0
arrow==1.3.0
asgiref==3.8.1
asteval==0.9.33
astropy==6.1.1
astropy-iers-data==0.2024.6.24.0.31.11
astroquery==0.4.7
asttokens==2.4.1
async-timeout==4.0.3
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
blake3==0.4.1
blinker==1.8.2
blosc2==2.7.0
bqplot==0.12.43
branca==0.7.2
Brotli==1.1.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
configparser==7.0.0
contourpy==1.2.1
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==42.0.8
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.10
dask==2024.6.2
decorator==5.1.1
Django==5.0.6
django-browser-reload==1.18.0
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-redis==5.4.0
django-tenants==3.6.1
django-timezone-field==7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
facebook-sdk==3.1.0
Faker==37.1.0
filelock==3.15.4
flower==2.0.1
fonttools==4.53.0
frozendict==2.4.4
fsspec==2024.6.1
future==1.0.0
gevent==24.10.2
greenlet==3.1.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
habanero==1.2.6
hitran-api==1.2.2.2
holidays==0.51
html5lib==1.1
httplib2==0.22.0
httptools==0.6.1
humanize==4.9.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
ipydatawidgets==4.3.5
ipyleaflet==0.19.1
ipython==8.25.0
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.11.1
ipyvuetify==1.9.4
ipywebrtc==0.6.0
ipywidgets==8.1.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jupyter-leaflet==0.19.1
jupyterlab_widgets==3.0.11
keyring==25.2.1
kiwisolver==1.4.5
kombu==5.3.7
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
ndindex==1.8
numexpr==2.10.1
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.22
pydyf==0.10.0
pyerfa==2.0.1.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyphen==0.15.0
pypng==0.20220715.0
python-barcode==0.15.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pythreejs==2.4.2
pytz==2024.1
pyvo==1.5.2
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.7
requests==2.32.3
rich==13.7.1
SecretStorage==3.3.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tailwind==3.1.5b0
text-unidecode==1.3
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
traittypes==0.2.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
widgetsnbextension==4.0.11
wkhtmltopdf==0.2
xyzservices==2024.6.0
zipp==3.19.2
zope.event==5.0
zope.interface==7.1.0
zopfli==0.2.3
